fix: :bug: Minor changes and fixes

+++ Userland/SampleCodeModule/tests/test_util.c
@@ -67,7 +67,7 @@ void endless_loop() {
 }
 
 void endless_loop_print(uint64_t wait) {
-	int64_t pid = sys_get_pid();
+	int32_t pid = sys_get_pid();
 
 	while (1) {
 		printf("%d ", pid);
